requirements.txt
Django==2.1.3
djangorestframework==3.9.0
django-filters==0.2.1
django-filter==2.0.0
factory-boy==2.11.1
Faker==0.9.2
